requirements.txt
rumps>=0.4.0
requests>=2.31.0
pycryptodome>=3.20.0
pyserial>=3.5
